# documentation/static/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Provider-neutral HTTP and server-sent-events API for running agent turns, reading retained messages, resolving tool decisions, and managing subagents.",
        "title": "AgentCLI HTTP API",
        "contact": {
            "name": "AgentCLI API support"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/sessions/{mainAgentSessionID}/subagent-confirmations": {
            "get": {
                "description": "Returns durable pending subagent confirmation requests so clients can recover after attaching late or reconnecting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Confirmations"
                ],
                "summary": "List pending confirmations for a main agent session",
                "operationId": "listPendingSubagentConfirmations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PendingSubagentConfirmationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/sessions/{mainAgentSessionID}/subagent-permissions": {
            "get": {
                "description": "Returns durable pending subagent permission requests so clients can recover after attaching late or reconnecting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "List pending permissions for a main agent session",
                "operationId": "listPendingSubagentPermissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Main agent session ID",
                        "name": "mainAgentSessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentcli.PendingSubagentPermissionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/sessions/{sessionID}/events": {
            "get": {
                "description": "Replays and follows every main-agent turn in a session, including queued lifecycle records, response-scope boundaries, and turns created automatically from subagent results. The session cursor is independent from each runtime event sequence.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Event streams"
                ],
                "summary": "Stream retained and live session activity",
                "operationId": "streamSessionEvents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resume after this session cursor",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Resume after this session cursor when after is omitted",
                        "name": "Last-Event-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One SSE data payload; the HTTP response remains open for more records",
                        "schema": {
                            "$ref": "#/definitions/agentcli.SessionEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agentcli.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "agentcli.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "agentcli.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/agentcli.APIError"
                }
            }
        },
...
        "agentcli.ConfirmationDecisionResponseValue": {
            "type": "object",
            "properties": {
                "answer": {
                    "$ref": "#/definitions/confirmation.Answer"
                },
                "call_id": {
                    "type": "string"
                },
                "confirmation_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
        "agentcli.ConfirmationRequestResponse": {
            "type": "object",
            "properties": {
                "call_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tool_name": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
...
        "agentcli.DecisionResponse": {
            "type": "object",
            "properties": {
                "call_id": {
                    "type": "string"
                },
                "permission_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/permission.DecisionType"
                }
            }
        },
        "agentcli.EventResponse": {
            "type": "object",
            "properties": {
                "confirmation": {
                    "$ref": "#/definitions/agentcli.ConfirmationRequestResponse"
                },
                "confirmation_decision": {
                    "$ref": "#/definitions/agentcli.ConfirmationDecisionResponseValue"
                },
                "decision": {
                    "$ref": "#/definitions/agentcli.DecisionResponse"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/agentcli.MessageResponse"
                },
                "permission": {
                    "$ref": "#/definitions/agentcli.PermissionRequestResponse"
                },
                "permission_mode": {
                    "$ref": "#/definitions/agentcli.PermissionModeChangeResponse"
                },
                "provider_event": {
                    "$ref": "#/definitions/agentcli.ProviderEventResponse"
                },
                "reason": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/agentcli.RunResultResponse"
                },
                "sequence": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "tool_request": {
                    "$ref": "#/definitions/agentcli.ToolRequestResponse"
                },
                "tool_result": {
                    "$ref": "#/definitions/agentcli.ToolResultEnvelopeResponse"
                },
                "turn_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/agentruntime.EventType"
                }
            }
        },
...
        "agentcli.MessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reasoning": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.ToolCallResponse"
                    }
                },
                "tool_result": {
                    "$ref": "#/definitions/agentcli.ToolResultResponse"
                },
                "turn_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/storage.MessageType"
                }
            }
        },
...
        "agentcli.PendingSubagentConfirmationsResponse": {
            "type": "object",
            "properties": {
                "confirmations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.SubagentConfirmationReference"
                    }
                }
            }
        },
        "agentcli.PendingSubagentPermissionsResponse": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agentcli.SubagentPermissionReference"
                    }
                }
            }
        },
...
        "agentcli.PermissionModeChangeResponse": {
            "type": "object",
            "properties": {
                "current": {
                    "$ref": "#/definitions/permission.Mode"
                },
                "previous": {
                    "$ref": "#/definitions/permission.Mode"
                }
            }
        },
...
        "agentcli.PermissionRequestResponse": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/permission.Action"
                    }
                },
                "call_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "risk": {
                    "$ref": "#/definitions/permission.Risk"
                },
                "session_id": {
                    "type": "string"
                },
                "tool_name": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
...
        "agentcli.ScopeEventResponse": {
            "type": "object",
            "properties": {
                "occurred_at": {
                    "type": "string"
                },
                "scope_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "subagent_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tool_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trigger_turn_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "pre_end_scope",
                        "end_scope"
                    ]
                }
            }
        },
...
        "agentcli.ServerTurnSource": {
            "type": "string",
            "enum": [
                "user",
                "subagent_result",
                "subagent_confirmation",
                "subagent_permission",
                "subagent_lifecycle"
            ],
            "x-enum-varnames": [
                "ServerTurnSourceUser",
                "ServerTurnSourceSubagentResult",
                "ServerTurnSourceSubagentConfirmation",
                "ServerTurnSourceSubagentPermission",
                "ServerTurnSourceSubagentLifecycle"
            ]
        },
        "agentcli.SessionActivityType": {
            "type": "string",
            "enum": [
                "turn_queued",
                "turn_admitted",
                "turn_cancelled",
                "turn_rejected",
                "turn_event",
                "scope_event",
                "subagent_confirmation",
                "subagent_permission",
                "subagent_closed"
            ],
            "x-enum-varnames": [
                "SessionActivityTurnQueued",
                "SessionActivityTurnAdmitted",
                "SessionActivityTurnCancelled",
                "SessionActivityTurnRejected",
                "SessionActivityTurnEvent",
                "SessionActivityScopeEvent",
                "SessionActivitySubagentConfirmation",
                "SessionActivitySubagentPermission",
                "SessionActivitySubagentClosed"
            ]
        },
        "agentcli.SessionEventResponse": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "events_url": {
                    "type": "string"
                },
                "queue_position": {
                    "type": "integer"
                },
                "runtime_event": {
                    "$ref": "#/definitions/agentcli.EventResponse"
                },
                "scope_event": {
                    "$ref": "#/definitions/agentcli.ScopeEventResponse"
                },
                "session_id": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/agentcli.ServerTurnSource"
                },
                "subagent_closed": {
                    "$ref": "#/definitions/agentcli.SubagentClosedReference"
                },
                "subagent_confirmation": {
                    "$ref": "#/definitions/agentcli.SubagentConfirmationReference"
                },
                "subagent_permission": {
                    "$ref": "#/definitions/agentcli.SubagentPermissionReference"
                },
                "subagent_result": {
                    "$ref": "#/definitions/agentcli.SubagentResultReference"
                },
                "turn_id": {
                    "type": "string"
                },
                "turn_url": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/agentcli.SessionActivityType"
                }
            }
        },
...
        "agentcli.SubagentClosedReference": {
            "type": "object",
            "properties": {
                "automatic": {
                    "type": "boolean"
                },
                "dropped_messages": {
                    "type": "integer"
                },
                "interrupted": {
                    "type": "boolean"
                },
                "previous_result_status": {
                    "$ref": "#/definitions/storage.SubagentResultStatus"
                },
                "previous_status": {
                    "$ref": "#/definitions/storage.SubagentStatus"
                },
                "subagent": {
                    "$ref": "#/definitions/agentcli.SubagentResponse"
                }
            }
        },
        "agentcli.SubagentConfirmationEventType": {
            "type": "string",
            "enum": [
                "requested",
                "resolved",
                "cancelled",
                "expired"
            ],
            "x-enum-varnames": [
                "SubagentConfirmationRequested",
                "SubagentConfirmationResolved",
                "SubagentConfirmationCancelled",
                "SubagentConfirmationExpired"
            ]
        },
        "agentcli.SubagentConfirmationReference": {
            "type": "object",
            "properties": {
                "confirmation": {
                    "$ref": "#/definitions/agentcli.ConfirmationRequestResponse"
                },
                "decision": {
                    "$ref": "#/definitions/agentcli.ConfirmationDecisionResponseValue"
                },
                "definition_name": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "subagent_id": {
                    "type": "string"
                },
                "subagent_session_id": {
                    "type": "string"
                },
                "subagent_turn_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/agentcli.SubagentConfirmationEventType"
                }
            }
        },
...
        "agentcli.SubagentPermissionEventType": {
            "type": "string",
            "enum": [
                "requested",
                "resolved",
                "cancelled",
                "expired"
            ],
            "x-enum-varnames": [
                "SubagentPermissionRequested",
                "SubagentPermissionResolved",
                "SubagentPermissionCancelled",
                "SubagentPermissionExpired"
            ]
        },
        "agentcli.SubagentPermissionReference": {
            "type": "object",
            "properties": {
                "decision": {
                    "$ref": "#/definitions/agentcli.DecisionResponse"
                },
                "definition_name": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/agentcli.PermissionRequestResponse"
                },
                "subagent_id": {
                    "type": "string"
                },
                "subagent_session_id": {
                    "type": "string"
                },
                "subagent_turn_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/agentcli.SubagentPermissionEventType"
                }
            }
        },
...
        "agentcli.SubagentResultReference": {
            "type": "object",
            "properties": {
                "definition_name": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "main_agent_session_id": {
                    "type": "string"
                },
                "main_agent_turn_id": {
                    "type": "string"
                },
                "next_step": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/agentcli.SubagentResultStatus"
                },
                "subagent_id": {
                    "type": "string"
                },
                "subagent_session_id": {
                    "type": "string"
                },
                "subagent_turn_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
...
        "agentcli.ToolRequestResponse": {
            "type": "object",
            "properties": {
                "call": {
                    "$ref": "#/definitions/agentcli.ToolCallResponse"
                },
                "session_id": {
                    "type": "string"
                },
                "turn_id": {
                    "type": "string"
                }
            }
        },
...
    }
}