# server_subagent_permissions.go
package agentcli

import (
	"net/http"

	"github.com/labstack/echo/v4"
)

func (server *Server) forwardSubagentPermissions(events <-chan SubagentPermissionEvent) {
	for {
		select {
		case <-server.context.Done():
			return
		case event, open := <-events:
			if !open {
				return
			}
			server.sessionEvents.publish(SessionEventResponse{
				Type:               SessionActivitySubagentPermission,
				Source:             ServerTurnSourceSubagentPermission,
				SessionID:          event.MainAgentSessionID,
				TurnID:             event.MainAgentTurnID,
				SubagentPermission: newSubagentPermissionReference(event),
			})
		}
	}
}

func newSubagentPermissionReference(event SubagentPermissionEvent) *SubagentPermissionReference {
	reference := &SubagentPermissionReference{
		Type: event.Type, SubagentID: event.SubagentID, DisplayName: event.DisplayName,
		DefinitionName: event.DefinitionName, SubagentSessionID: event.SubagentSessionID, SubagentTurnID: event.SubagentTurnID,
	}
	if event.Request != nil {
		request := newPermissionRequestResponse(*event.Request)
		reference.Permission = &request
	}
	if event.Decision != nil {
		decision := newDecisionResponse(*event.Decision)
		reference.Decision = &decision
	}
	return reference
}

// listPendingSubagentPermissions godoc
// @Summary List pending permissions for a main agent session
// @ID listPendingSubagentPermissions
// @Description Returns durable pending subagent permission requests so clients can recover after attaching late or reconnecting.
// @Tags Permissions
// @Produce json
// @Param mainAgentSessionID path string true "Main agent session ID"
// @Success 200 {object} PendingSubagentPermissionsResponse
// @Failure 400 {object} APIErrorResponse
// @Router /v1/sessions/{mainAgentSessionID}/subagent-permissions [get]
func (server *Server) listPendingSubagentPermissions(c echo.Context) error {
	mainAgentSessionID := c.Param("mainAgentSessionID")
	if mainAgentSessionID == "" {
		return writeAPIError(c, http.StatusBadRequest, "invalid_request", "main agent session ID is required")
	}
	events, err := server.agent.PendingSubagentPermissions(c.Request().Context(), mainAgentSessionID)
	if err != nil {
		return server.writeRuntimeError(c, err)
	}
	response := PendingSubagentPermissionsResponse{Permissions: make([]SubagentPermissionReference, 0, len(events))}
	for _, event := range events {
		response.Permissions = append(response.Permissions, *newSubagentPermissionReference(event))
	}
	return writeJSON(c, http.StatusOK, response)
}